requirements.txt
absl-py==2.1.0
ale-py==0.8.1
anyio==4.4.0
box2d-py==2.3.5
certifi==2024.6.2
charset-normalizer==3.3.2
clip==1.0
cloudpickle==3.0.0
cmake==3.29.5.1
contourpy==1.1.1
cycler==0.12.1
Cython==0.29.37
decorator==4.4.2
diskcache==5.6.3
etils==1.3.0
exceptiongroup==1.2.1
Farama-Notifications==0.0.4
filelock==3.14.0
fonttools==4.53.0
fsspec==2024.6.0
ftfy==6.2.0
glfw==2.7.0
gymnasium==0.29.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.4
idna==3.7
imageio==2.34.1
imageio-ffmpeg==0.5.1
importlib-metadata==7.1.0
importlib-resources==6.4.0
jax==0.4.13
jaxlib==0.4.13
jinja2==3.1.4
kiwisolver==1.4.5
lit==18.1.7
llama-cpp-python==0.2.79
lz4==4.3.3
MarkupSafe==2.1.5
matplotlib==3.7.5
ml-dtypes==0.2.0
moviepy==1.0.3
mpmath==1.3.0
mujoco==3.1.6
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu11==11.7.91
ollama==0.2.1
open-clip-torch==2.24.0
opencv-python==4.10.0.82
opt-einsum==3.3.0
packaging==24.1
pandas==2.0.3
pillow==10.3.0
proglog==0.1.10
protobuf==5.27.2
pygame==2.5.2
PyOpenGL==3.1.7
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
safetensors==0.4.3
scipy==1.10.1
sentencepiece==0.2.0
Shimmy==0.2.1
six==1.16.0
sniffio==1.3.1
stable-baselines3==2.3.2
swig==4.2.1
sympy==1.12.1
timm==1.0.7
tokenizers==0.19.1
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tqdm==4.66.4
transformers==4.42.1
triton==2.0.0
typing-extensions==4.12.2
tzdata==2024.1
urllib3==2.2.1
wcwidth==0.2.13
zipp==3.19.2
